requirements.txt
streamlit==1.24.0
scikit-learn>=1.0.2
nltk>=3.6.7
pandas>=1.4.2
numpy>=1.22.3
matplotlib>=3.5.2
seaborn>=0.11.2
ntscraper>=0.3.0
pillow>=9.0.0
wordcloud>=1.8.1